# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global College LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sections:
    get:
      operationId: listSections
      tags: [sections]
      summary: List sections for a course
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      tags: [sections]
      summary: Create a section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionBody"
      responses:
        "201":
          description: Section created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /sections/{id}:
    get:
      operationId: getSection
      tags: [sections]
      summary: Get section by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    put:
      operationId: updateSection
      tags: [sections]
      summary: Update section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionBody"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /settings:
    get:
      operationId: listSettings
      tags: [settings]
      summary: List all settings
      responses:
        "200":
          description: List of settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
    put:
      operationId: updateSettingsBulk
      tags: [settings]
      summary: Update settings in bulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
  /settings/{key}:
    put:
      operationId: updateSetting
      tags: [settings]
      summary: Update a single setting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
              required: [value]
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [settings]
      summary: List recent announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [settings]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sentBy:
                  type: integer
                title:
                  type: string
                message:
                  type: string
                targetType:
                  type: string
                targetId:
                  type: integer
              required: [sentBy, title, message]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Admin verify/approve payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        userName:
          type: string
        courseName:
          type: string
        amount:
          type: number
        method:
          type: string
          enum: [card, bank_transfer, easypaisa, jazzcash, cash]
        status:
          type: string
          enum: [pending, verified, rejected]
        receiptUrl:
          type: string
        createdAt:
          type: string
      required: [id, userId, courseId, amount, method, status, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        courseId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [card, bank_transfer, easypaisa, jazzcash, cash]
        receiptUrl:
          type: string
      required: [courseId, amount, method]
    VerifyPaymentBody:
      type: object
      properties:
        status:
          type: string
          enum: [verified, rejected]
        notes:
          type: string
      required: [status]
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        studentName:
          type: string
        courseName:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatar:
          type: string
        createdAt:
          type: string
      required: [id, studentName, courseName, content, rating, createdAt]
    CreateTestimonialBody:
      type: object
      properties:
        studentName:
          type: string
        courseName:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatar:
          type: string
      required: [studentName, courseName, content, rating]
...
    Section:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        order:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, courseId, title, order]
    CreateSectionBody:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        order:
          type: integer
      required: [courseId, title]
...
    Setting:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        label:
          type: string
        category:
          type: string
      required: [id, key, value]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        sentBy:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        title:
          type: string
        message:
          type: string
        sentAt:
          type: string
      required: [id, sentBy, title, message, sentAt]